refactor expression package
